Organize - common directory for shared code - removed unused imports
