+++ src/components/infopopups/CrossAppealPopup.js
@@ -1,6 +1,5 @@
 import React, { Component }from 'react';
-import './infopopup.css';
-import Help from './Help';
+import Help from './common/Help';
 
 class CrossAppealPopup extends Component {
 
@@ -49,7 +48,11 @@ class CrossAppealPopup extends Component {
                     Cross Appeal (If Applicable)
                 </div>
                 <div className="info-modal-content">
-                    <div className="info-modal-primary-heading ">Were you served with any of the following documents?</div>
+                    <div className="info-modal-primary-heading row">
+                        <div className="col-lg-offset-1 col-md-11 col-sm-11 col-lg-11 col-md-offset-1 col-sm-offset-1">
+                            Were you served with any of the following documents?
+                        </div>
+                    </div>
                     {sections}
                     <Help
                         URL="https://www.courtofappealbc.ca/appellant-guidebook"
