+++ src/components/infopopups/ReplyBookPopup.js
@@ -1,5 +1,5 @@
 import React, { Component } from 'react';
-import Help from "./Help";
+import Help from "./common/Help";
 
 class ReplyBookPopup extends Component {
