+++ src/components/infopopups/common/infopopup.css
